+++ crates/leto/src/lib.rs
@@ -27,8 +27,8 @@ pub use application::view::{ArrayView, ArrayViewMut};
 pub use application::{
     concat, covariance, median_all, median_axis, pad, pearson_correlation, quantile_all,
     quantile_axis, split, stack, Array1, Array2, Array3, ArrayView1, ArrayView2, ArrayView3,
-    ArrayViewMut1, ArrayViewMut2, ArrayViewMut3, AxisIter, AxisIterMut, Interpolation, PadWidth,
-    ScalarOperand,
+    ArrayViewMut1, ArrayViewMut2, ArrayViewMut3, AxisIter, AxisIterMut, ElementIter, IndexedIter,
+    Interpolation, PadWidth, ScalarOperand, Windows,
 };
 
 #[cfg(feature = "ndarray-compat")]